requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
async-timeout==5.0.1
attrs==25.3.0
cassandra-driver==3.29.2
click==8.2.1
colorama==0.4.6
discord==2.3.2
discord.py==2.5.2
frozenlist==1.7.0
geomet==0.2.1.post1
idna==3.10
multidict==6.6.3
propcache==0.3.2
python-dotenv==1.1.1
pytz==2025.2
six==1.17.0
typing_extensions==4.14.1
yarl==1.20.1
